% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_colors.R
\name{mcolor}
\alias{mcolor}
\title{Get a Color from Material Design Color Palette}
\usage{
mcolor(color, shade = NULL)
}
\arguments{
\item{color}{A character string specifying the color name. It can be a regular color (e.g.,
\code{"red"}, \code{"blue"}, \code{"green"}), or an accent color (e.g., \code{"red_accent"}, \code{"blue_accent"}). Refer to the list
below for available color options.}

\item{shade}{A numeric value specifying the shade of the color.
The default value is \code{500} for the regular color and \code{200} for the accent color.}
}
\value{
A character string representing the hex code of the selected color, or \code{NA} if the color is invalid.
}
\description{
This function provides a color from the Material Design color palette.
You can select from a range of regular colors and accent colors, and specify a particular shade.
If the provided color or shade is invalid, the function will return \code{NA} with a warning.
}
\details{
The Material Design color palette includes various shades for each color. Most colors have shades
from \code{100} to \code{900} in increments of one hundred, plus the color \code{50}. The smaller the number, the lighter the color, and the greater
the number, the darker the color. Accent colors, such as \code{red_accent} or \code{blue_accent}, only have shades
of \code{100}, \code{200}, \code{400}, and \code{700}.

Available colors in the Material Design palette are:

Regular colors:
\itemize{
\item \code{red}, \code{pink}, \code{purple}, \code{deep_purple}, \code{indigo}, \code{blue}, \code{light_blue}, \code{cyan}, \code{teal}, \code{green}, \code{light_green}, \code{lime}, \code{yellow}, \code{amber}, \code{orange}, \code{deep_orange}, \verb{brown, grey}, \code{blue_grey}
}

Accent colors:
\itemize{
\item \code{red_accent}, \code{pink_accent}, \code{purple_accent}, \code{deep_purple_accent}, \code{indigo_accent}, \code{blue_accent}, \code{light_blue_accent}, \code{cyan_accent}, \code{teal_accent}, \code{green_accent}, \code{light_green_accent}, \code{lime_accent}, \code{yellow_accent}, \code{amber_accent}, \code{orange_accent}, \code{deep_orange_accent}
}

For regular colors, valid shades are \code{50}, \code{100}, \code{200}, \code{300}, \code{400}, \code{500}, \code{600}, \code{700}, \code{800}, and \code{900}.
For accent colors, valid shades are \code{100}, \code{200}, \code{400}, and \code{700}.
}
\examples{
mcolor("blue", 500) # Returns hex code for blue color, shade 500
mcolor("blue_accent", 200) # Returns hex code for blue accent, shade 200
mcolor("red", 1000) # Returns NA and prints a warning

}
